requirements.txt
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.0rc3
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy==1.5.1
mypy-extensions==1.0.0
psycopg2==2.9.7
PySocks==1.7.1
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
requests==2.31.0
requests-file==1.5.1
simple-websocket==0.10.1
six==1.16.0
SQLAlchemy==2.0.20
tldextract==3.6.0
typing_extensions==4.7.1
urllib3==2.0.5
Werkzeug==2.3.7
wsproto==1.2.0
zope.event==5.0
zope.interface==6.0
